backend/requirements.txt
ultralytics
pyzmq
opencv-python
numpy
fastapi
uvicorn[standard]
websockets
pyserial
python-dotenv
customtkinter
pyttsx3
